+++ src/app/fetcher/index.test.js
@@ -8,7 +8,7 @@ import fetch from './index.js';
 const __dirname = path.dirname(new URL(import.meta.url).pathname);
 
 const { expect } = chai;
-const facebookTermsHTML = fs.readFileSync(path.resolve(__dirname, '../../test/fixtures/facebook_terms_raw.html'), { encoding: 'utf8' });
+const facebookTermsHTML = fs.readFileSync(path.resolve(__dirname, '../../../test/fixtures/facebook_terms_raw.html'), { encoding: 'utf8' });
 
 nock('https://www.facebook.com', {
   reqheaders: { 'Accept-Language': 'en' }
