+++ src/app/history/index.test.js
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import chai from 'chai';
 
-import { resetGitRepository, gitSnapshot, gitVersion } from '../../test/helper.js';
+import { resetGitRepository, gitSnapshot, gitVersion } from '../../../test/helper.js';
 import { SNAPSHOTS_PATH, VERSIONS_PATH, recordSnapshot, recordVersion, recordRefilter } from './index.js';
 
 const { expect } = chai;
