+++ src/app/loader/index.test.js
@@ -5,7 +5,7 @@ import chai from 'chai';
 import loadServiceDeclarations from './index.js';
 
 const __dirname = path.dirname(new URL(import.meta.url).pathname);
-const SERVICE_DECLARATIONS_PATH = path.resolve(__dirname, '../..', config.get('serviceDeclarationsPath'));
+const SERVICE_DECLARATIONS_PATH = path.resolve(__dirname, '../../..', config.get('serviceDeclarationsPath'));
 
 const { expect } = chai;
